feat: HRT architecture v2 — blitz mode, 6-regime options, Ralph loop, PostgreSQL-only
Major changes across the entire system:

- CLAUDE.md + prompts: rewrite loop orchestration to use Ralph Wiggum --file flag,
  eliminating bash wrapper scripts (start_*.sh deleted, run_*.sh added as thin wrappers)
- Prompts: major expansions to research_loop, trading_loop, research_options (6-regime
  QQQ framework), research_shared (breadth-first tool mandate), and new ARCHITECTURE.md
- Agents: daily-planner added; execution-researcher, ml-scientist, quant-researcher,
  risk, strategy-rd all updated with refined mandates and tool catalogs
- MCP tools: blitz mode execution paths, alpaca broker integration, full options
  execution wiring, new _base/_tool_def modules, and _impl/execution/ml expansions
- DB layer: db.py overhauled with connection pooling, health checks, and context managers
- Data: acquisition_pipeline, economic_storage, options_storage, pg_storage all
  cleaned up post-DuckDB removal; universe.py simplified
- Execution: alpaca_broker.py added; broker_factory, order_lifecycle, paper_broker updated
- Learning/monitoring: metrics.py expanded, reflection.py refactored, skill_tracker fixed
- Tests: 30+ new test files covering MCP tools, integration (blitz mode), and unit layers
- Memory: AMZN/META/NVDA ticker memories added; QQQ/SPY/XLE updated with 6-regime data
- Docs: blitz mode implementation docs and options analyst summary added

+++ src/quantstack/audit/decision_log.py
@@ -100,7 +100,7 @@ class DecisionLog:
                     agent_role              VARCHAR NOT NULL,
                     symbol                  VARCHAR,
                     action                  VARCHAR,
-                    confidence              DOUBLE,
+                    confidence              DOUBLE PRECISION,
                     input_context_hash      VARCHAR,
                     market_data_snapshot    JSON,
                     portfolio_snapshot      JSON,
